backend/requirements.txt
boto3==1.26.29
botocore==1.29.29
click==8.1.3
Flask==2.2.2
Flask-CLI==0.4.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-dotenv==0.21.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
Werkzeug==2.2.2
